bugfixes_1

+++ sel_repeat.h
@@ -15,7 +15,7 @@
 
 int init_serv(int port, char* hostname);
 int connect_rdt(int port, char* hostname);
-int read_sr(int meta_i, void *buf, size_t nbyte);
-int write_sr(int meta_i, void *buf, size_t count);
+int read_sr(int meta_i, void *buf, unsigned int nbyte);
+int write_sr(int meta_i, void *buf, unsigned int count);
 
 #endif
